+++ include/skyerrors.h
@@ -1,0 +1,9 @@
+
+#ifndef SKY_ERRORS_H
+#define SKY_ERRORS_H
+
+#define SKY_OK            0
+#define SKY_ERROR         0xFFFFFFFF
+
+#endif
+